chore: isolate `dispatch` mod
